[Rename] 프로젝트를 Lobby로 변경, App name을 socket으로 변경
