--- backend/chat_server/chatting/chat/admin.py
@@ -1,6 +1,0 @@
-from django.contrib import admin
-from .models import User, Room
-
-
-admin.site.register(User)
-admin.site.register(Room)
