+++ backend/lobby_server/config/urls.py
@@ -17,6 +17,5 @@ from django.contrib import admin
 from django.urls import include, path
 
 urlpatterns = [
-    path('single/', include('chatting.chat.urls')),
     path('admin/', admin.site.urls),
 ]
