+++ backend/lobby_server/lobby/socket/admin.py
@@ -1,0 +1,1 @@
+from django.contrib import admin
