+++ backend/lobby_server/lobby/socket/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class ChatConfig(AppConfig):
+class socketConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'chatting.chat'
+    name = 'lobby.socket'
